+++ source/slang/slang.vcxproj
@@ -171,7 +171,6 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\slang.h" />
-    <ClInclude Include="bytecode.h" />
     <ClInclude Include="compiler.h" />
     <ClInclude Include="core.meta.slang.h" />
     <ClInclude Include="decl-defs.h" />
@@ -222,10 +221,8 @@
     <ClInclude Include="type-system-shared.h" />
     <ClInclude Include="val-defs.h" />
     <ClInclude Include="visitor.h" />
-    <ClInclude Include="vm.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="bytecode.cpp" />
     <ClCompile Include="check.cpp" />
     <ClCompile Include="compiler.cpp" />
     <ClCompile Include="diagnostics.cpp" />
@@ -262,7 +259,6 @@
     <ClCompile Include="token.cpp" />
     <ClCompile Include="type-layout.cpp" />
     <ClCompile Include="type-system-shared.cpp" />
-    <ClCompile Include="vm.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="slang.natvis" />
