+++ source/slang/slang.vcxproj.filters
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="Header Files">
@@ -10,9 +10,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\slang.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="bytecode.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="compiler.h">
@@ -165,14 +162,8 @@
     <ClInclude Include="visitor.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="vm.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="bytecode.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="check.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -279,9 +270,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="type-system-shared.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="vm.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
