--- source/slang/vm.h
@@ -1,19 +1,0 @@
-// vm.h
-#ifndef SLANG_VM_H_INCLUDED
-#define SLANG_VM_H_INCLUDED
-
-// This file defines a virtual machine for executing
-// code that has been converted to Slang bytecode.
-//
-
-#include "../core/basic.h"
-
-namespace Slang
-{
-
-
-
-}
-
-
-#endif
